requirements.txt
django 
psycopg2 
psycopg2-binary 
django-admin-autocomplete-filter  
django-import-export
django_s3_storage
django-s3-sqlite
zappa
